fix: move components from sync page
